Fixes Issue #25, renamed packages with respective changes to the build.xml
