+++ javadoc/edu/ucsb/cs56/projects/languages/vocab_quiz/class-use/ForeignVocabWord.html
@@ -1,0 +1,149 @@
+<!DOCTYPE HTML PUBLIC "-//W3C//DTD HTML 4.01 Transitional//EN" "http://www.w3.org/TR/html4/loose.dtd">
+<!-- NewPage -->
+<html lang="en">
+<head>
+<!-- Generated by javadoc (1.8.0_111) on Mon Oct 31 10:53:26 PDT 2016 -->
+<title>Uses of Class edu.ucsb.cs56.projects.languages.vocab_quiz.ForeignVocabWord</title>
+<meta name="date" content="2016-10-31">
+<link rel="stylesheet" type="text/css" href="../../../../../../../stylesheet.css" title="Style">
+<script type="text/javascript" src="../../../../../../../script.js"></script>
+</head>
+<body>
+<script type="text/javascript"><!--
+    try {
+        if (location.href.indexOf('is-external=true') == -1) {
+            parent.document.title="Uses of Class edu.ucsb.cs56.projects.languages.vocab_quiz.ForeignVocabWord";
+        }
+    }
+    catch(err) {
+    }
+//-->
+</script>
+<noscript>
+<div>JavaScript is disabled on your browser.</div>
+</noscript>
+<!-- ========= START OF TOP NAVBAR ======= -->
+<div class="topNav"><a name="navbar.top">
+<!--   -->
+</a>
+<div class="skipNav"><a href="#skip.navbar.top" title="Skip navigation links">Skip navigation links</a></div>
+<a name="navbar.top.firstrow">
+<!--   -->
+</a>
+<ul class="navList" title="Navigation">
+<li><a href="../../../../../../../edu/ucsb/cs56/projects/languages/vocab_quiz/package-summary.html">Package</a></li>
+<li><a href="../../../../../../../edu/ucsb/cs56/projects/languages/vocab_quiz/ForeignVocabWord.html" title="class in edu.ucsb.cs56.projects.languages.vocab_quiz">Class</a></li>
+<li class="navBarCell1Rev">Use</li>
+<li><a href="../../../../../../../overview-tree.html">Tree</a></li>
+<li><a href="../../../../../../../deprecated-list.html">Deprecated</a></li>
+<li><a href="../../../../../../../index-all.html">Index</a></li>
+<li><a href="../../../../../../../help-doc.html">Help</a></li>
+</ul>
+</div>
+<div class="subNav">
+<ul class="navList">
+<li>Prev</li>
+<li>Next</li>
+</ul>
+<ul class="navList">
+<li><a href="../../../../../../../index.html?edu/ucsb/cs56/projects/languages/vocab_quiz/class-use/ForeignVocabWord.html" target="_top">Frames</a></li>
+<li><a href="ForeignVocabWord.html" target="_top">No&nbsp;Frames</a></li>
+</ul>
+<ul class="navList" id="allclasses_navbar_top">
+<li><a href="../../../../../../../allclasses-noframe.html">All&nbsp;Classes</a></li>
+</ul>
+<div>
+<script type="text/javascript"><!--
+  allClassesLink = document.getElementById("allclasses_navbar_top");
+  if(window==top) {
+    allClassesLink.style.display = "block";
+  }
+  else {
+    allClassesLink.style.display = "none";
+  }
+  //-->
+</script>
+</div>
+<a name="skip.navbar.top">
+<!--   -->
+</a></div>
+<!-- ========= END OF TOP NAVBAR ========= -->
+<div class="header">
+<h2 title="Uses of Class edu.ucsb.cs56.projects.languages.vocab_quiz.ForeignVocabWord" class="title">Uses of Class<br>edu.ucsb.cs56.projects.languages.vocab_quiz.ForeignVocabWord</h2>
+</div>
+<div class="classUseContainer">
+<ul class="blockList">
+<li class="blockList">
+<ul class="blockList">
+<li class="blockList"><a name="edu.ucsb.cs56.projects.languages.vocab_quiz">
+<!--   -->
+</a>
+<h3>Uses of <a href="../../../../../../../edu/ucsb/cs56/projects/languages/vocab_quiz/ForeignVocabWord.html" title="class in edu.ucsb.cs56.projects.languages.vocab_quiz">ForeignVocabWord</a> in <a href="../../../../../../../edu/ucsb/cs56/projects/languages/vocab_quiz/package-summary.html">edu.ucsb.cs56.projects.languages.vocab_quiz</a></h3>
+<table class="useSummary" border="0" cellpadding="3" cellspacing="0" summary="Use table, listing methods, and an explanation">
+<caption><span>Method parameters in <a href="../../../../../../../edu/ucsb/cs56/projects/languages/vocab_quiz/package-summary.html">edu.ucsb.cs56.projects.languages.vocab_quiz</a> with type arguments of type <a href="../../../../../../../edu/ucsb/cs56/projects/languages/vocab_quiz/ForeignVocabWord.html" title="class in edu.ucsb.cs56.projects.languages.vocab_quiz">ForeignVocabWord</a></span><span class="tabEnd">&nbsp;</span></caption>
+<tr>
+<th class="colFirst" scope="col">Modifier and Type</th>
+<th class="colLast" scope="col">Method and Description</th>
+</tr>
+<tbody>
+<tr class="altColor">
+<td class="colFirst"><code>int</code></td>
+<td class="colLast"><span class="typeNameLabel">ForeignVocabQuiz.</span><code><span class="memberNameLink"><a href="../../../../../../../edu/ucsb/cs56/projects/languages/vocab_quiz/ForeignVocabQuiz.html#randomIndex-java.util.ArrayList-">randomIndex</a></span>(java.util.ArrayList&lt;<a href="../../../../../../../edu/ucsb/cs56/projects/languages/vocab_quiz/ForeignVocabWord.html" title="class in edu.ucsb.cs56.projects.languages.vocab_quiz">ForeignVocabWord</a>&gt;&nbsp;aList)</code>
+<div class="block">Get a random index from <code>vocabList</code>.</div>
+</td>
+</tr>
+</tbody>
+</table>
+</li>
+</ul>
+</li>
+</ul>
+</div>
+<!-- ======= START OF BOTTOM NAVBAR ====== -->
+<div class="bottomNav"><a name="navbar.bottom">
+<!--   -->
+</a>
+<div class="skipNav"><a href="#skip.navbar.bottom" title="Skip navigation links">Skip navigation links</a></div>
+<a name="navbar.bottom.firstrow">
+<!--   -->
+</a>
+<ul class="navList" title="Navigation">
+<li><a href="../../../../../../../edu/ucsb/cs56/projects/languages/vocab_quiz/package-summary.html">Package</a></li>
+<li><a href="../../../../../../../edu/ucsb/cs56/projects/languages/vocab_quiz/ForeignVocabWord.html" title="class in edu.ucsb.cs56.projects.languages.vocab_quiz">Class</a></li>
+<li class="navBarCell1Rev">Use</li>
+<li><a href="../../../../../../../overview-tree.html">Tree</a></li>
+<li><a href="../../../../../../../deprecated-list.html">Deprecated</a></li>
+<li><a href="../../../../../../../index-all.html">Index</a></li>
+<li><a href="../../../../../../../help-doc.html">Help</a></li>
+</ul>
+</div>
+<div class="subNav">
+<ul class="navList">
+<li>Prev</li>
+<li>Next</li>
+</ul>
+<ul class="navList">
+<li><a href="../../../../../../../index.html?edu/ucsb/cs56/projects/languages/vocab_quiz/class-use/ForeignVocabWord.html" target="_top">Frames</a></li>
+<li><a href="ForeignVocabWord.html" target="_top">No&nbsp;Frames</a></li>
+</ul>
+<ul class="navList" id="allclasses_navbar_bottom">
+<li><a href="../../../../../../../allclasses-noframe.html">All&nbsp;Classes</a></li>
+</ul>
+<div>
+<script type="text/javascript"><!--
+  allClassesLink = document.getElementById("allclasses_navbar_bottom");
+  if(window==top) {
+    allClassesLink.style.display = "block";
+  }
+  else {
+    allClassesLink.style.display = "none";
+  }
+  //-->
+</script>
+</div>
+<a name="skip.navbar.bottom">
+<!--   -->
+</a></div>
+<!-- ======== END OF BOTTOM NAVBAR ======= -->
+</body>
+</html>
