+++ javadoc/edu/ucsb/cs56/projects/languages/vocab_quiz/package-tree.html
@@ -1,0 +1,136 @@
+<!DOCTYPE HTML PUBLIC "-//W3C//DTD HTML 4.01 Transitional//EN" "http://www.w3.org/TR/html4/loose.dtd">
+<!-- NewPage -->
+<html lang="en">
+<head>
+<!-- Generated by javadoc (1.8.0_111) on Mon Oct 31 10:53:26 PDT 2016 -->
+<title>edu.ucsb.cs56.projects.languages.vocab_quiz Class Hierarchy</title>
+<meta name="date" content="2016-10-31">
+<link rel="stylesheet" type="text/css" href="../../../../../../stylesheet.css" title="Style">
+<script type="text/javascript" src="../../../../../../script.js"></script>
+</head>
+<body>
+<script type="text/javascript"><!--
+    try {
+        if (location.href.indexOf('is-external=true') == -1) {
+            parent.document.title="edu.ucsb.cs56.projects.languages.vocab_quiz Class Hierarchy";
+        }
+    }
+    catch(err) {
+    }
+//-->
+</script>
+<noscript>
+<div>JavaScript is disabled on your browser.</div>
+</noscript>
+<!-- ========= START OF TOP NAVBAR ======= -->
+<div class="topNav"><a name="navbar.top">
+<!--   -->
+</a>
+<div class="skipNav"><a href="#skip.navbar.top" title="Skip navigation links">Skip navigation links</a></div>
+<a name="navbar.top.firstrow">
+<!--   -->
+</a>
+<ul class="navList" title="Navigation">
+<li><a href="../../../../../../edu/ucsb/cs56/projects/languages/vocab_quiz/package-summary.html">Package</a></li>
+<li>Class</li>
+<li>Use</li>
+<li class="navBarCell1Rev">Tree</li>
+<li><a href="../../../../../../deprecated-list.html">Deprecated</a></li>
+<li><a href="../../../../../../index-all.html">Index</a></li>
+<li><a href="../../../../../../help-doc.html">Help</a></li>
+</ul>
+</div>
+<div class="subNav">
+<ul class="navList">
+<li>Prev</li>
+<li>Next</li>
+</ul>
+<ul class="navList">
+<li><a href="../../../../../../index.html?edu/ucsb/cs56/projects/languages/vocab_quiz/package-tree.html" target="_top">Frames</a></li>
+<li><a href="package-tree.html" target="_top">No&nbsp;Frames</a></li>
+</ul>
+<ul class="navList" id="allclasses_navbar_top">
+<li><a href="../../../../../../allclasses-noframe.html">All&nbsp;Classes</a></li>
+</ul>
+<div>
+<script type="text/javascript"><!--
+  allClassesLink = document.getElementById("allclasses_navbar_top");
+  if(window==top) {
+    allClassesLink.style.display = "block";
+  }
+  else {
+    allClassesLink.style.display = "none";
+  }
+  //-->
+</script>
+</div>
+<a name="skip.navbar.top">
+<!--   -->
+</a></div>
+<!-- ========= END OF TOP NAVBAR ========= -->
+<div class="header">
+<h1 class="title">Hierarchy For Package edu.ucsb.cs56.projects.languages.vocab_quiz</h1>
+</div>
+<div class="contentContainer">
+<h2 title="Class Hierarchy">Class Hierarchy</h2>
+<ul>
+<li type="circle">java.lang.Object
+<ul>
+<li type="circle">edu.ucsb.cs56.projects.languages.vocab_quiz.<a href="../../../../../../edu/ucsb/cs56/projects/languages/vocab_quiz/FileRead.html" title="class in edu.ucsb.cs56.projects.languages.vocab_quiz"><span class="typeNameLink">FileRead</span></a></li>
+<li type="circle">edu.ucsb.cs56.projects.languages.vocab_quiz.<a href="../../../../../../edu/ucsb/cs56/projects/languages/vocab_quiz/ForeignVocabOutput.html" title="class in edu.ucsb.cs56.projects.languages.vocab_quiz"><span class="typeNameLink">ForeignVocabOutput</span></a></li>
+<li type="circle">edu.ucsb.cs56.projects.languages.vocab_quiz.<a href="../../../../../../edu/ucsb/cs56/projects/languages/vocab_quiz/ForeignVocabQuiz.html" title="class in edu.ucsb.cs56.projects.languages.vocab_quiz"><span class="typeNameLink">ForeignVocabQuiz</span></a></li>
+<li type="circle">edu.ucsb.cs56.projects.languages.vocab_quiz.<a href="../../../../../../edu/ucsb/cs56/projects/languages/vocab_quiz/ForeignVocabQuizGUI.html" title="class in edu.ucsb.cs56.projects.languages.vocab_quiz"><span class="typeNameLink">ForeignVocabQuizGUI</span></a></li>
+<li type="circle">edu.ucsb.cs56.projects.languages.vocab_quiz.<a href="../../../../../../edu/ucsb/cs56/projects/languages/vocab_quiz/ForeignVocabWord.html" title="class in edu.ucsb.cs56.projects.languages.vocab_quiz"><span class="typeNameLink">ForeignVocabWord</span></a></li>
+<li type="circle">edu.ucsb.cs56.projects.languages.vocab_quiz.<a href="../../../../../../edu/ucsb/cs56/projects/languages/vocab_quiz/ForeignVocabWordTest.html" title="class in edu.ucsb.cs56.projects.languages.vocab_quiz"><span class="typeNameLink">ForeignVocabWordTest</span></a></li>
+</ul>
+</li>
+</ul>
+</div>
+<!-- ======= START OF BOTTOM NAVBAR ====== -->
+<div class="bottomNav"><a name="navbar.bottom">
+<!--   -->
+</a>
+<div class="skipNav"><a href="#skip.navbar.bottom" title="Skip navigation links">Skip navigation links</a></div>
+<a name="navbar.bottom.firstrow">
+<!--   -->
+</a>
+<ul class="navList" title="Navigation">
+<li><a href="../../../../../../edu/ucsb/cs56/projects/languages/vocab_quiz/package-summary.html">Package</a></li>
+<li>Class</li>
+<li>Use</li>
+<li class="navBarCell1Rev">Tree</li>
+<li><a href="../../../../../../deprecated-list.html">Deprecated</a></li>
+<li><a href="../../../../../../index-all.html">Index</a></li>
+<li><a href="../../../../../../help-doc.html">Help</a></li>
+</ul>
+</div>
+<div class="subNav">
+<ul class="navList">
+<li>Prev</li>
+<li>Next</li>
+</ul>
+<ul class="navList">
+<li><a href="../../../../../../index.html?edu/ucsb/cs56/projects/languages/vocab_quiz/package-tree.html" target="_top">Frames</a></li>
+<li><a href="package-tree.html" target="_top">No&nbsp;Frames</a></li>
+</ul>
+<ul class="navList" id="allclasses_navbar_bottom">
+<li><a href="../../../../../../allclasses-noframe.html">All&nbsp;Classes</a></li>
+</ul>
+<div>
+<script type="text/javascript"><!--
+  allClassesLink = document.getElementById("allclasses_navbar_bottom");
+  if(window==top) {
+    allClassesLink.style.display = "block";
+  }
+  else {
+    allClassesLink.style.display = "none";
+  }
+  //-->
+</script>
+</div>
+<a name="skip.navbar.bottom">
+<!--   -->
+</a></div>
+<!-- ======== END OF BOTTOM NAVBAR ======= -->
+</body>
+</html>
