+++ src/edu/ucsb/cs56/projects/languages/vocab_quiz/FileRead.java
@@ -1,4 +1,4 @@
-package edu.ucsb.cs56.W16.yvalencia.foreignvocabquiz;
+package edu.ucsb.cs56.projects.languages.vocab_quiz;
 import java.awt.*;
 import javax.swing.*;
 import java.io.*;
