+++ src/edu/ucsb/cs56/projects/languages/vocab_quiz/ForeignVocabQuiz.java
@@ -1,4 +1,4 @@
-package edu.ucsb.cs56.W16.yvalencia.foreignvocabquiz;
+package edu.ucsb.cs56.projects.languages.vocab_quiz;
 
 import java.io.*;
 import java.util.*;
