+++ src/edu/ucsb/cs56/projects/languages/vocab_quiz/images/wrong.gif
